Reorganized TwitchIRC script.
