+++ Runtime/TwitchIRC.ReceiveThread.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: edf19b1fe691d406e8571572f139cef8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
