+++ Runtime/TwitchIRC.SendThread.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 37c2c487fe1a943f09ad3d7e2eaddffd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
